pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-assistant"
version = "2.0.0"
description = "AI-powered Gmail inbox management — direct API, model-agnostic LLM gateway"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "google-api-python-client>=2.120.0",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "litellm>=1.30.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]
postgres = [
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
]

[project.scripts]
gmail-assistant = "src.main:cli_entry"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
